requirements.txt
pymupdf==1.25.3 
openai
pandas
python-dotenv
openpyxl